+++ lib/execute/execute.test.ts
@@ -141,8 +141,7 @@ describe("execute", () => {
     });
   });
 
-  // FIXME: Make FrameBytes work with IArrayConstructor?
-  describe.skip("byte.strings", () => {
+  describe("byte.strings", () => {
     it("reads n characters", () => {
       const input = "\\1\\a";
       const result = execute(input);